# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/reports": {
      "post": {
        "operationId": "ModerationController_report",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReportDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "moderation"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
...
    "/admin/users": {
      "get": {
        "operationId": "AdminController_listUsers",
        "parameters": [
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/admin/users/{id}": {
      "get": {
        "operationId": "AdminController_getUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/admin/users/{id}/status": {
      "post": {
        "operationId": "AdminController_setStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetStatusDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "admin"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/admin/users/{id}/impersonate": {
      "post": {
        "operationId": "AdminController_impersonate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "admin"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/admin/reports": {
      "get": {
        "operationId": "AdminController_reports",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/admin/reports/{id}/resolve": {
      "post": {
        "operationId": "AdminController_resolveReport",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveReportDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "admin"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/admin/audit": {
      "get": {
        "operationId": "AdminController_audit",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/admin/settings": {
      "get": {
        "operationId": "AdminController_settings",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/admin/settings/{key}": {
      "put": {
        "operationId": "AdminController_setSetting",
        "parameters": [
          {
            "name": "key",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetSettingDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "dateChain API",
    "description": "REST API for the dateChain dating platform",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateReportDto": {
        "type": "object",
        "properties": {
          "reportedId": {
            "type": "string",
            "format": "uuid"
          },
          "category": {
            "type": "string",
            "enum": [
              "SPAM",
              "ABUSE",
              "FAKE",
              "INAPPROPRIATE",
              "UNDERAGE",
              "OTHER"
            ]
          },
          "reason": {
            "type": "string"
          }
        },
        "required": [
          "reportedId",
          "category"
        ]
      },
...
      "SetStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "PAUSED",
              "BANNED",
              "DELETED"
            ]
          }
        },
        "required": [
          "status"
        ]
      },
      "ResolveReportDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "RESOLVED",
              "DISMISSED"
            ]
          },
          "resolution": {
            "type": "string"
          },
          "ban": {
            "type": "boolean"
          }
        },
        "required": [
          "status"
        ]
      },
      "SetSettingDto": {
        "type": "object",
        "properties": {
          "value": {
            "type": "object"
          }
        },
        "required": [
          "value"
        ]
      }
    }
  }
}